+++ axelor-production/src/main/java/com/axelor/apps/production/service/OperationOrderWorkflowService.java
@@ -29,6 +29,7 @@ import com.axelor.apps.production.db.repo.ManufOrderRepository;
 import com.axelor.apps.production.db.repo.OperationOrderDurationRepository;
 import com.axelor.apps.production.db.repo.OperationOrderRepository;
 import com.axelor.apps.production.db.repo.ProductionConfigRepository;
+import com.axelor.apps.production.exceptions.IExceptionMessage;
 import com.axelor.apps.production.service.app.AppProductionService;
 import com.axelor.apps.production.service.config.ProductionConfigService;
 import com.axelor.apps.stock.db.StockMove;
@@ -36,6 +37,8 @@ import com.axelor.apps.stock.service.StockMoveService;
 import com.axelor.apps.tool.date.DurationTool;
 import com.axelor.auth.AuthUtils;
 import com.axelor.exception.AxelorException;
+import com.axelor.exception.db.repo.TraceBackRepository;
+import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
@@ -427,7 +430,8 @@ public class OperationOrderWorkflowService {
     return operationOrder;
   }
 
-  public LocalDateTime computePlannedEndDateT(OperationOrder operationOrder) {
+  public LocalDateTime computePlannedEndDateT(OperationOrder operationOrder)
+      throws AxelorException {
 
     if (operationOrder.getWorkCenter() != null) {
       return operationOrder
@@ -441,7 +445,8 @@ public class OperationOrderWorkflowService {
     return operationOrder.getPlannedStartDateT();
   }
 
-  public long computeEntireCycleDuration(OperationOrder operationOrder, BigDecimal qty) {
+  public long computeEntireCycleDuration(OperationOrder operationOrder, BigDecimal qty)
+      throws AxelorException {
 
     long machineDuration = this.computeMachineDuration(operationOrder, qty);
 
@@ -454,7 +459,8 @@ public class OperationOrderWorkflowService {
     }
   }
 
-  public long computeMachineDuration(OperationOrder operationOrder, BigDecimal qty) {
+  public long computeMachineDuration(OperationOrder operationOrder, BigDecimal qty)
+      throws AxelorException {
     ProdProcessLine prodProcessLine = operationOrder.getProdProcessLine();
     WorkCenter workCenter = prodProcessLine.getWorkCenter();
 
@@ -465,6 +471,13 @@ public class OperationOrderWorkflowService {
     if (workCenterTypeSelect == IWorkCenter.WORK_CENTER_MACHINE
         || workCenterTypeSelect == IWorkCenter.WORK_CENTER_BOTH) {
       Machine machine = workCenter.getMachine();
+      if (machine == null) {
+        throw new AxelorException(
+            workCenter,
+            TraceBackRepository.CATEGORY_MISSING_FIELD,
+            I18n.get(IExceptionMessage.WORKCENTER_NO_MACHINE),
+            workCenter.getName());
+      }
       duration += machine.getStartingDuration();
 
       BigDecimal durationPerCycle = new BigDecimal(prodProcessLine.getDurationPerCycle());